Fixed data overflow issue when int64 value is interpreted as float/double.

+++ modules/core/src/persistence.cpp
@@ -2410,7 +2410,7 @@ FileNode::operator float() const
 
     if( type == INT )
     {
-        return (float)readInt(p);
+        return (float)readLong(p);
     }
     else if( type == REAL )
     {
@@ -2431,7 +2431,7 @@ FileNode::operator double() const
 
     if( type == INT )
     {
-        return (double)readInt(p);
+        return (double)readLong(p);
     }
     else if( type == REAL )
     {
@@ -2442,6 +2442,7 @@ FileNode::operator double() const
 }
 
 double FileNode::real() const  { return double(*this); }
+
 std::string FileNode::string() const
 {
     const uchar* p = ptr();
